remove xplat react regex test cases
Summary:
The xplat react regex test cases added in D66147456. But the xplat react regex
in configerator and the test cases in here might be out of sync. We recently added
a CI time verifier for the xplat react regex patterns, and added test cases in
configerator (D66590020), so user can modify the regex and its test cases in one diff
in configerator.

Reviewed By: RajivTS

Differential Revision: D67254117

fbshipit-source-id: 6857b2c5dd347b39adc18626589f612fda1029e0