escape redirection path in eden config toml on Windows
Summary:
S477425 happened because when we add a new redirection on windows, we didn't escape the backslashes on windows. (Look at the test plan before this change)
The root cause was in this fact that we didn't have any serialization for `redirection` and `redirection_targets` in `CheckoutConfig`
Then the file path save wrong in the toml file.
To fix this issue, we need to consider two things:
1- This diff adds a `serialize_path_map` serialization to both `redirection` and `redirection_targets`
2- Also, when we want to read path from the `config.toml` file, we should consider that all the \\ should convert to \ in addition to / to \
The changes on deserialized function consider these changes.
It is explained in the test plan that the redirect works fine and eden can successfully restart by these changes

Reviewed By: jdelliot

Differential Revision: D67293753

fbshipit-source-id: d27239c4d76d619a29e785bd4378d62b820216d5