pathhistory: fix fetching bug when a path prefix was not a directory
Summary:
When a prefix of the given path was a file (or symlink) in the commit history, the load_tree_entry() will fail,
since it was trying to read a directory.  For example, let's assume the given path is `a/b/c`, and `a/b` is a
directory now; but in the commit history, `a/b` was not a directory (regular file or symlink).

This diff fixes the issue by ignoring the non-directory items, this will make pathhistory ignore those commits
that the prefix was not a directory. If the file was a symlink to another folder, it will not follow the link.

Reviewed By: muirdm

Differential Revision: D67412803

fbshipit-source-id: 5e8ace29e2b26d295f31d3014cb1c67cf98c10e5