eden-client refactoring: move Thrift calls from minitop into Eden instance
Summary:
## Context
The stack of diffs aims to encapsulate Thrift calls from EdenFS cli commands/subcommands into the eden instance. This is done to improve the overall design and maintainability of the code. The overall purpose of the stack of diffs is to improve the code structure and make it easier to maintain and modify in the future.

## This Diff

Moved thrift calls from minitop subcommand to be Eden instance methods. The API of the methods accept an optional connected thrift client so we can save one async call if the caller already has a client ready.

Reviewed By: jdelliot

Differential Revision: D67403258

fbshipit-source-id: fca1ce1f06b3b49e42acec1a9f0907f883470157