eden-client refactoring: move simple Thrift calls into Eden instance
Summary:
## Context

The stack of diffs aims to encapsulate Thrift calls from EdenFS cli commands/subcommands into the eden instance. This is done to improve the overall design and maintainability of the code. The overall purpose of the stack of diffs is to improve the code structure and make it easier to maintain and modify in the future.

## This Diff

- Refactored the client connect code into a method to avoid repeated copy/paste: `get_connected_thrift_client`
- Moved Thrift calls out of these commands: `gc`, `prefetch_profile`, `debug`

Reviewed By: jdelliot

Differential Revision: D67401738

fbshipit-source-id: 5c3c79362946ba38651e1b8f566a6da93b2bda60