Ensure expired and non-existing bubbles don't result in 500 status code responses
Summary:
Due to incorrect error handling, we get a lot of 500s in Eden API for requests that are invalid to begin with. Either they refer to an non-existent bubble or they refer to a bubble that has already expired.
{F1974017446}

Which leads to logs like above in our scuba table. This diff makes the necessary changes to ensure that we return such cases with 4xx error codes so they don't get classified as server errors

Reviewed By: lmvasquezg

Differential Revision: D67458467

fbshipit-source-id: ff5ae4c8c48a260a5b8d4c218156f30821212788