Fix flaky Gitimport integration test
Summary:
The `test-gitimport-with-non-standard-file-mode.t` test was created to test an error condition and how gitimport behaves in such a case. However, since this test was created we added retries to every error in gitimport to provide more resilience to the gitimport process. Owning to those retries, sometimes the expected error message in this test ends up including a retry statement like `retrying attempt 2 of 4...` which ends up making the test flaky.

This diff modifies the test to print only the top 6 lines so we don't even allow the retry line to get included as part of the test output.

Reviewed By: YousefSalama

Differential Revision: D67522222

fbshipit-source-id: f71097061157ce7ddde67fb52f4b18be9dc0c50d