fix missing fetchblob and fetch tree counters from getCounter
Summary:
When we setup OBC API for `SaplingBackingStore` fetch duration counters, we keep OBC API under an eden config `isOBCEnabled_`, and in the else condition we record `SaplingBackingStoreStats::fetchBlob` and `SaplingBackingStoreStats::fetchTree` FB303 counters. That is wrong for two reasons:
1- the fetchBlob and fetchTree FB303 counters are general value for all repos while the OBC API are repo specific value. Therefore these two counters doesn't have overlap and we need both in some cases.
2- On the hosts that we enable OBC API, we will miss the value of `fetch_blob_us` and `fetch_tree_us` counters from thrift getCounter functions. Because OBC API values doesn't show up in the thrift call.

It is better to remove the else part and collect FB303 counters regardless of the OBC API.
See test plan on a devserver that `isOBCEnabled_` was True.

Reviewed By: jdelliot

Differential Revision: D67550926

fbshipit-source-id: 670bc2b71052a8e6d7be6512fb91af010e24f946