Stop using the legacy addStatValue() for eden rss counter
Summary:
Stop using the legacy `addStatValue()` call that checks to see if it needs to re-export counters each time it is used.
It doesn't worth exporting this as a timeseries.  We mainly only care about the last 60-second timeseries level. FBAgent runs every 60-second and reports this value every 60-second to ODS.
In the current code, we only update this once every 30 seconds then we are basically just reporting an average of the last 2 data points which doesn't have any value.

Reviewed By: jdelliot

Differential Revision: D67613587

fbshipit-source-id: ffa7d347c1066971809600950e92a3a3b4a3d63a