absorb: drop unused applyFileStackEdits
Summary:
`applyFileStackEdits` was the initial idea that the `AbsorbEdit` and `wdir()`
state lives outside the main stack, stack revs are integers.

The latest design, `applyFileStackEditsWithAbsorbId`, (see also D67525914 and
D67525963) mantains absorb edits directly in linelog, and `wdir()` is part of
the main stack, state revs can be non-integers. This new design replaces the
old. So let's just delete the old logic as it is not going to be used.

Reviewed By: muirdm

Differential Revision: D67611454

fbshipit-source-id: 53f1e6b3e6381b9f2a0b036092853544590c2617