linelog: comment that Rev does not have to be an integer
Summary:
I found that it'd be easier for absorb to just operate at the line log level,
instead of dealing with the diff chunks. For example, let the blame of a.txt
be:

  Rev 1 : 111
  wdir(): 122
  Rev 1 : 222
  Rev 2 : 333
  wdir(): 444
  Rev 2 : 222

We can use Rev 3.1, Rev 3.2 to represent the wdir changes:

  Rev 1   : 111
  Rev 3.1 : 122
  Rev 1   : 222
  Rev 2   : 333
  Rev 3.2 : 444
  Rev 2   : 222

and we can just replace 3.1 with 1.1 to "absorb" it in rev 1 (by checking out
1.99), replace 3.2 with 2.2 to absorb it to rev 2 (by checking out 3.99):

  Rev 1   : 111
  Rev 1.1 : 122
  Rev 1   : 222
  Rev 2   : 333
  Rev 2.2 : 444
  Rev 2   : 222

If the user wants to move "122" to rev 2, instead of rev 1, we can replace 1.1
with 2.1 to achieve that.

Note the above example is a bit simplified as it does not demostrate how
deletion works. Practically, the rev numbers are not just stored on the LINEs
but also JGE, JL instructions. The `remapRevs` API takes care of all these
instructions so deletion should just work.

It's also worth noting that `0.1`, `0.2` cannot be losslessly represented in
binary. Practically, we should choose numbers like `0.0009765625`, `1/(1<<10)`
instead.

Reviewed By: zzl0

Differential Revision: D67525914

fbshipit-source-id: af4394d89908f1fcdc800d8f96c51e49d2f078e1