fix or allow elided_named_lifetimes warning
Summary:
There are a couple of places where we were relying on elided lifetimes implicity matching with an already-named lifetime.  This is now a warning, so let's stop doing that.

For `hgproto::sshproto::request`, this comes from the use of `nom` macros, so suppress the warning there.

For `derived_data_manager::manager::derive`, we were matching on the lifetime added by `async_recursion`, so we need to explicitly name the lifetime to get around this.

Reviewed By: mitrandir77, RajivTS

Differential Revision: D67735986

fbshipit-source-id: da58202d48859bd0238704e8e8ce9e8a84093ddf