Add suffix support for Notifications
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff

This diff introduces the concepts of included/excluded Suffixes to the getChangesV2 pipeline.

The files returned from the API will always match any defined included suffixes, and never have any excluded suffixes. If a suffix is both included and excluded, then it is excluded.

A suffix is defined as the substring that ends the string. As currently defined, a suffix could include content past the file extention, including the , eg for file.ext1.ext2, ex2, ext1.ext2, and file.ext1.ext2 are valid suffixes

# Technical Details
This diff creates the thrift definitions for the included/excluded suffix types

# Discussion Points

Reviewed By: jdelliot

Differential Revision: D67613986

fbshipit-source-id: f18957a490d3095324d46046f57233bba92565aa