Make access logs optional, defaulting to false
Summary:
Access logs are very rarely useful, but on a routine basis they just crowd the output, making anything else invisible. The underlying data is in Scuba anyway, so nothing is lost.

The motivation for this is actually AWS, so we have less data to send back to prod (thus making e.g. log rotation less of a priority).

Reviewed By: clara-9

Differential Revision: D67792653

fbshipit-source-id: 46b6b4ba4cedf133a567583984c3cd884ce8485d