fix FinishedMount::populate() to log backing_store_type
Summary: This field was added to `FinishedMount` but never added to `populate()`. This diff fixes this so we log the `backing_store_type` to scuba. The motivation behind sending this diff was that I tried to use this data before (to see the distribution of types) and it was missing.

Reviewed By: jdelliot

Differential Revision: D67767064

fbshipit-source-id: 6cc9e4d3872f0d2adf04f15f8edf1c9aac078264