Improve warning when not all files are fetched
Summary:
if viewing a commit with >1000 files, the pages for 1001-1500 and beyond would show no files and would say "Showing files 1001 – 1000 out of 1600 total" which is wrong.

This could happen due to a limit in the number of files we fetch for commits for performance. We can fix the "1001 – 1000"  thing by using the totalFile count instead of the fetched files.

Additionally, when we detect this case, we can use stronger wording to explain that not all files have been fetched.

Reviewed By: sggutier

Differential Revision: D67809455

fbshipit-source-id: 5c235da8ad7f41496bdcaec4c88333b56df0592d