Don't show <DiffStats> for public commits changed files
Summary: The <DiffStats> shows the LOC for a commit when viewing it. This should be largely irrelevant for public commits, which you can only incidentally view by having nothing selected when you're checked out on a public commit. The previous diff stopped showing changed files by default, and this feature is disabled for similar reasons.

Reviewed By: sggutier

Differential Revision: D67809454

fbshipit-source-id: a99873f5b08060d82368f62599e2c8fe7262e804