Remove commit changed files fetch limit
Summary:
We had a limit of 1000 files on changed files in a commit. This was not handled nicely, as the UI would let you try to page beyond the first 1000, then just show nothing.

The reason for the limit is performance. You may be on a commit with 100,000 changed files. The sl log command will dutifully fetch all those files and report them to the UI.

The performance concern is two-fold:
1. fetching many files from sl may take a long time and use machine resources
2. having the UI deal with lots of files in memory is visually laggy

However, the limit as-is is actually not really helping us out much with these issues. The limit is applied *after* the results are fetched from sl. This means the full expensive fetch is always done. We already paginate files, so we don't render all the files even if we did fetch them. The only possible perf improvement this makes is that we don't need to send the message from the server containing the list of all the files.

I considered having a two-fetch system, which would load like 5000 files, then if you try to click "next page", it would load the full list. But this will do the expensive `sl log` call twice. We might as well just load all the files right away if you're looking at a commit.

The really bad perf situations are typically from a large public change you happen to be checked out on that you didn't author. So we can leave the limit in place for public commits. In a follow up diff, we can make a perf improvement for this in how we fetch changes, and also make the UI not show this poorly when paging through files.

Reviewed By: sggutier

Differential Revision: D67809457

fbshipit-source-id: 77db271f2b1bd5fd3e9e4be99eafe1bcd5b4b8f8