hooks: Fix bug with scuba logging
Summary:
We used to mingle all samples so that they would report the same hook name.

This was due to the fact that we were modifying the adding the entry for the hook name before cloning scuba, so we would override's the last sample's hook name.

This was a regression introduced in D66177243. The bug had existed before and been fixed with D64901151.

We love stateful programming.

Reviewed By: RajivTS

Differential Revision: D67898955

fbshipit-source-id: e32477e96e0511076ab406cf9588c8293d0946c6