Log timing metrics for Gitimport part of Mononoke Git push to Scuba
Summary: Similar to previous diff, this diff logs the timing metrics for the Gitimport part of the push process since its one of the most expensive components of the push workflow. The gitimport part consists of 4 major sub-components, and the scuba logging added in this diff will help us identify which one needs the most improvement.

Reviewed By: YousefSalama, lmvasquezg

Differential Revision: D67901127

fbshipit-source-id: 1f55a00b8f63ad991d868c375881333afcdd4c3d