Fix eden health check exit code
Summary:
**Background**:
Previously, when detecting issues with the Eden setup, we would send an exit code of 1 to identify the issue on the VS Code side. However, VS Code's execution logic would interpret this exit code as a command execution failure, causing it to terminate the code flow prematurely and preventing the reporting of Eden-related issues to the end user.

**Changes**:
To address this issue, we have introduced a custom isExitError method within VS Code's execution logic. This method ensures that an exit code of 1 is treated as a non-error scenario, allowing the code flow to continue uninterrupted when Eden-related issues are detected.
With this change, VS Code will no longer terminate prematurely due to Eden issues, enabling more robust error handling and improved user experience.

Reviewed By: genevievehelsel

Differential Revision: D67868950

fbshipit-source-id: 5948e692df4f4e447834c63b413d9b3e503373d1