stackEdit: construct Dag from commit stack
Summary:
This will be used in the absorb UI. Note we contruct a new Dag instead of using
a subset of the existing "main" dag so it is free from race conditions (ex.
click absorb, repo changed, before refreshing the main dag), and can work with
a different commit stack (ex. after re-ordering).

Reviewed By: zzl0

Differential Revision: D67878992

fbshipit-source-id: f54b6056673e96ed7e4e5662a2bd6caaeb13ea29