stackEdit: maybeBuildFileStacks -> useFileStack
Summary:
`maybeBuildFileStacks` does not ensure that the file stacks are actually used.
Using it puts the `stack` in a "weird" state that absorb edits cannot be applied,
because the absorb edits are updated in the linelog, but the stack is not using
linelog as source of truth.

Make the `maybeBuildFileStacks` API private and use `useFileStack` to ensure a
consistent state.

Reviewed By: evangrayk

Differential Revision: D67881523

fbshipit-source-id: b117991c9a40cd14cb4a2027b6b9fe52ce9ce50b