stackEdit: use branded type for FileRev
Summary:
Previously, a `number` can be used for both the revision of a commit stack, or
a file stack, which is error prone. See D67927194 for an example.

This diff makes `FileRev` ("rev" used by the file stack) a "branded" TypeScript
type so functions taking `FileRev` no longer accepts regular `number`.
Hopefully this can make the type cheker more useful to detect related mistakes.

That said, the "branded" type is not perfect, namely:
- Verbose casting. For example, when `x` is `FileRev`, both `x + 1` and
  `Math.max(x, x)` become just `number` and requires `as FileRev`.
  Constants like `0` requires `as FileRev` casting too.

Note this diff only makes `FileRev` branded, the commit `Rev` is not branded
yet (due to complexity). To make the mistake in D67927194 fail typecheck, we
need `Rev` to be branded too so TypeScript can report error. See:

  const commitRev = 1 as (number & {__brand: 'commit'});
  const fileRev = 2 as (number & {__brand: 'file'});
  const anyRev = 3 as number;
  // No error
  anyRev === fileRev;
  // Error: This comparison appears to be unintentional because the types 'number & { __brand: "commit"; }' and 'number & { __brand: "file"; }' have no overlap.
  commitRev === fileRev;

Ideally, TypeScript can provide more native support for this use-case, to
reduce verbosity, and be optionally more strict.

Reviewed By: evangrayk

Differential Revision: D67980848

fbshipit-source-id: 4f466bae44377d4bbbcc61c771569f7a383d0813