absorb: tweak the instruction text
Summary:
The current instruction sounds like Drag-and-Drop has to be done to select
amend destinations. Actually, the destinations are pre-selected and only needed
to be adjusted. Make it clear about the pre-selection.

Since the text becomes long, I made the "omitted" message only appear when
omitting actually happens. I also added a message if there is no valid absorb
destinations.

Reviewed By: evangrayk

Differential Revision: D67985670

fbshipit-source-id: d1a8a73feff0d082ef6f4a8b3bc61287b0c3ea3d