call fbinit for oss build
Summary:
Call facebookInit for the buck built OSS build as well as the full-fat facebook build. fbinit is required for the stats/ods Rust crates, and the eden integration tests use the OSS build for build speed reasons (and some integration tests want to test that counters are incremented properly).

I acknowledge that it doesn't "make sense" for the OSS build to call fbinit, but I think this is just a symptom of the original "doesn't make sense" of using the OSS build of EdenFS for tests.

Reviewed By: MichaelCuevas

Differential Revision: D67726336

fbshipit-source-id: a14b463a82fe82853880b6a6e539296bcca4888b