Update EjecaError to be a custom type with custom toString
Summary:
We were creating `EjecaError`s via `new Error` and a type cast, and assigning extra properties. But this means `.toString` would not print the additional fields like exit code and stdout.

This makes it hard to debug issues, as we don't see stdout/stderr in the logs (though users should still see in their output UI).

This error format might be visible in the UI for certain errors, though I think that's fine?

I also noticed the EjecaError type is missing `code`, but some callers still check that. We should fix that too (not in this diff)

Reviewed By: muirdm

Differential Revision: D67994455

fbshipit-source-id: f7b2933187c17800037716609120f76417f2cdb6