Fix for pushing to empty repo on Mononoke Git server
Summary:
In Protocol V1, the capability of the server is advertised along with the ref advertisement. This is done by null-separating the capabilities from the first ref while the rest of the refs are sent normally.

However, since the capabilities are tied to ref advertisement in the write path, having no refs mean that we won't send any capabilities to the client. This is what was wrong with the current implementation. Now the logic has been corrected to make it consistent with [the spec](https://git-scm.com/docs/pack-protocol#_reference_discovery)

Reviewed By: andreacampi

Differential Revision: D67978800

fbshipit-source-id: bb9ea76dca8f27ac3ed1f4b6b06fc55f795e4804