modern sync: bail on changeset upload failure
Summary: So i was seeing erros like: upload contents succeeds, filenodes success, trees succeeds and... changeset fails. There are multiple root causes for this that are fixed already but I basically want (for now) to avoid a situation where we keep uploading stuff  and then it becomes a pain to know the last full changeset that was synced, once  smarter retries are implemented I'll let that happen again.

Reviewed By: andreacampi

Differential Revision: D68015789

fbshipit-source-id: ab49b48c4efb6d62c03cd6f32aedcb8719a13e95