goto: swap redacted content for human readable message
Summary:
When checking out files, if the content matches the redacted tombstone, swap the contents to a human readable message. This logic was missing from the Rust checkout.

I considered adding this logic deeper in the scmstore FileStore object, but I didn't like that since it would make it hard for us to see the "real" content served by the server. Only performing the substitution for the working copy manifestation of the contents seems better.

Reviewed By: quark-zju, andreacampi

Differential Revision: D67958797

fbshipit-source-id: 73d6cddcee53530e562b5df543584a4f0f85fab0