testing: change getenv() to only use default on None
Summary:
The current `getenv(<name>, <default>)` uses the default string when <name> is not set or is set to empty string. This makes it impossible to tell the difference between not-set and empty-string.

Fix by only using the default value when <name> is not set. This is how os.getenv works, so I assume this was the original intention.

Note I also fixed another bug where `alt` was not being propagated to the `self.parent.getenv()` call.

For some env vars it is important to tell the difference between not-set and empty-string (e.g. $IFS).

Reviewed By: quark-zju

Differential Revision: D67992543

fbshipit-source-id: d0e09ad00d1b28275d421a84ff8db2a283085c41