backingstore: use nullptr for blob-not-found
Summary:
Use nullptr intead of exceptions to represent the low level "not found" state. liubov-dmitrieva discovered through profiling that we spend a lot of CPU time on exceptions. In particular, we first try a local-only fetch, and if the blob doesn't exist, we raise an exception and then enqueue the blob to be fetched in allow-remote mode.

Note that I only changed the fetch-single-blob path. The fetch batch path still uses exceptions. This is because, in practice, the batch path is used with "allow remote", so "not found" is unexpected.

I changed the low level SaplingNativeBackingStore::getBlob to use nullptr to mean not-found, and then updated callers. I wasn't sure the best way to do this across FFI/cxx, so I added a simple wrapper struct with a bool representing whether the blob was found or not. cxx doesn't seem to supports options yet.

Reviewed By: liubov-dmitrieva

Differential Revision: D67363553

fbshipit-source-id: 24cccb014730f178ad6c0765f3b7f94b61c82d82