backingstore: use nullptr for aux-not-found
Summary:
Use nullptr intead of exceptions to represent the low level "not found" state for file and tree aux data. liubov-dmitrieva discovered through profiling that we spend a lot of CPU time on exceptions. In particular, we first try a local-only fetch, and if the aux data doesn't exist, we raise an exception and then enqueue the aux data to be fetched in allow-remote mode.

I changed the low level SaplingNativeBackingStore::get{Blob,Tree}AuxData to use nullptr to mean not-found, and then updated callers.

Note that I did not change the batch aux codepaths. Since these typically fetch remotely, "not found" results are not expected (so it is okay to keep using exceptions).

Reviewed By: liubov-dmitrieva

Differential Revision: D67545966

fbshipit-source-id: 73e8fa7a2584192faa13295870bff6c38eec0727