backingstore: use nullptr for tree-not-found
Summary:
Use nullptr intead of exceptions to represent the low level "not found" state. liubov-dmitrieva discovered through profiling that we spend a lot of CPU time on exceptions. In particular, we first try a local-only fetch, and if the tree doesn't exist, we create an exception (in the folly::Try) and then enqueue the tree to be fetched in allow-remote mode.

Note that I only changed the fetch-single-tree path. The fetch batch path still uses exceptions. This is because, in practice, the batch path is used with "allow remote", so "not found" is unexpected.

I changed the low level SaplingNativeBackingStore::getTree to use nullptr to mean not-found, and then updated callers.

I also tweaked scmstore `single()` method to return `Ok(None)` if a local-only fetch found no data. This `None` is used by the backingstore to propagate `nullptr` instead of an exception.

Reviewed By: liubov-dmitrieva

Differential Revision: D67363549

fbshipit-source-id: 9184ac65f51f7008c3f293df6d2eb892a0f4599f