Minor fixes for Eden health-check
Summary:
Details:
1. **Exit Code = 255 for Exceptions**
We now exit with an exit code of 255 when Eden Health-Check encounters an exception during execution. This allows us to capture the exception details and print them to stdout, which can later be retrieved in VSCode for telemetry purposes.
2. **Fix Mount Parameters**
We've updated the way we send mount parameters to ensure they're sent as a list of strings instead of a single string. This fix resolves any issues related to incorrect parameter formatting.
3. **Remediation Command Execution without CONNECT_HOST**
In some cases, the CONNECT_HOST parameter may not correctly determine the target host for executing remediation commands. To resolve this issue, we've updated the executeCommand function to internally determine the target host, eliminating the need for the CONNECT_HOST parameter.
4. **Show Command Succeeded Notification**
We've added a notification that appears after a remediation command has successfully completed its execution. This notification is displayed within the then clause to ensure users are notified only when the command has run successfully.

Differential Revision: D68037504

fbshipit-source-id: 828a97e5a932d6254008fb6b2056f0aa599989f9