Integration test highlighting failure while pushing ref deltas in empty repo in Mononoke Git
Summary:
We saw this happening in two production repos so we do know that this is a thing. However, my every attempt to reproduce this in an integration test using git client failed miserably. mitrandir77 and I tried to have an integration test with the exact repo contents that were failing in production, but still it worked.

Giving up, I have manually reproduced this by making a CURL request to the Git server on the push endpoint which highlights the problem. This isn't ideal, but its better than nothing and it will help prevent regression in the future.

Differential Revision: D68016834

fbshipit-source-id: 4384d7b22d27d0a8cd4df66cbfdf3d4c711fa7b9