Implement concurrent derivation for GitDeltaManifest v2
Summary:
At Git push time (and also during GitImport), we try to derive all the git derived data types using their batch implementation. For types that have a hierarchical dependency, the batch implementation is essentially a serial for loop because we can't do any better. However, that's not the case for `GitDeltaManifestV2` since its derivation is independent of its predecessors, i.e. if A is the parent commit and B is the child commit, GDMv2 for both A and B can be derived parallely.

However, so far we were using just the sequential batch implementation thus missing out on an optimization opportunity. This diff takes care of that.

Differential Revision: D68100151

fbshipit-source-id: 9af241e19e61f006c696ccce8081e07337467d48