enabled inline comments with no code suggestions to be shown
Summary:
**Context**: explained by the title

**Solution**:
With code suggestion, we get commit hash from CodePatch so that we can compare code in editor with the commit code. We do this to realigned the line number in case of code change in vscode editor.

For entities without codepatch, we have to fetch the commit hash via 'changeset' in comment's data.  Hence the added graphql.

**PS.** In the next diff I will add in the ability to resolve a comment, so user can resolve and disband comments without codepatch.

Differential Revision: D67951639

fbshipit-source-id: c45be6a747653bab3707e1fde1c117015e3396ea