Implement suffix filtering to changeSince
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff

This diff implements the suffix filtering functionality. It does this by checking if the path

# Technical Details
Insert suffix filtering into the existing isPathIncluded function. Logic is the same as for roots except it uses suffixes. This is a very simple for loop matcher that may have performance issues if the number of suffixes grows large.

# Discussion Points
As we add more filtering, we put more strain on the current simple matcher implementation.

Benchmarking in a followup diff

ends with using str.compare vs suffix=str.substring benchmark
{F1974173247}

Reviewed By: jdelliot

Differential Revision: D67763311

fbshipit-source-id: af9f0bcb4fac3c3c463a30605b328264233ca557