Only show fetch from cloud button if commit cloud enabled
Summary:
Check if the commitcloud extension is enabled in order to show the fetch from cloud button after loading all commits.

I rearranged the fetch code into its own component so it can use Suspense to lazy load the commitcloud fetch.

I had to make a new atom to store this enabled state, as we only had an expensive check for cloud sync status available.

This new check means we include `extensions.commitcloud` in our sl config call, which unforutnately means we get all `extension.*` values... I'm thinking of updating our call to avoid this. We tried this in the past but it requires a newer version of Sapling, which we can't guarantee in OSS.

Reviewed By: nsblake

Differential Revision: D68041107

fbshipit-source-id: 4fdfd7d9e98d26e529b75977ca44c6af8f5e7e17