Implement support in gitimport for refs pointing to blobs and trees
Summary:
This enables `gitimport` to import repos that have refs pointing to blobs and trees. There are three surfaces where we need this support, `gitimport`, `remote-gitimport` and finally `Mononoke Git server`.

`Mononoke Git server` will also require different plumbing for push and pull path. Those will be addressed in follow up diffs.

NOTE: This diff does not address the scenario of ref pointing to a tree where none of the tree's children are imported in Mononoke. That will be addressed in the very next diff.

Differential Revision: D68103177

fbshipit-source-id: 3b7cb298f85ca1e405be0ff969f59199bfb6cbb7