Relax rules for git object upload method in storage layer
Summary:
The `upload_git_object` method in `gitimport` and `remote-gitimport` was supported by `upload_non_blob_git_object` method in the storage layer. In addition to the explicit naming, we had added a check in place to fail the method in case the object being uploaded was a git blob.

The reason for that was `gitimport` was in its initial stages and we wanted to ensure that everytime a git blob is uploaded, the correspdonding file changes are uploaded as Mononoke content blobs.

However, recently we realized that there can be refs pointing directly to blobs. In that case, there is no need for uploading the corresponding file changes in Mononoke since there is no bonsai commit to point to them. Thus, our constraint needs to be relaxed which is what's being done in this diff.

Reviewed By: singhsrb

Differential Revision: D68101616

fbshipit-source-id: a3137230fd603c2061c1af48e91e3e4e01483875