Refactor prompt for user confirmation
Summary:
## Context

This diff stack focuses on the support of removing multiple path with the Rust version of remove command.

In the current state machine design, the path validation and path type determination are parts of the state machine flow. This is inconvenient when we want to prompt the user to confirm the removal when multiple paths are given. Things are even more complicated when multiple paths are of different types: some paths are active eden checkouts while some are non-eden directories. (**metrics to be added to track this in the future**)

Python version of the remove command is not handling this case well. It will only prompt the user for the first non-eden path if any is given:
 {F1974226015}

## This Diff

We move the prompt outside of these states:
- ActiveEdenMount
- RegFile
- UnKnown

So we can prompt the user before we start the state machine.

## Nest Step

With the new method implemented, we can decouple the validation/determination steps from the state machine flow. The designed flow for multiple paths would be:
1. ~~For each path given, determine the type~~
2. ~~Prompt the user once we have all the paths determined~~
3. After confirmation is given, iterate through the paths and start the state machine for each of them

Reviewed By: MichaelCuevas

Differential Revision: D67920928

fbshipit-source-id: 43e0d7d718ff13e79aa70eb5b38be95456c0f0c3