Use function to replace State enum
Summary:
## Context

This diff stack focuses on the support of removing multiple path with the Rust version of remove command.

In the current state machine design, the path validation and path type determination are parts of the state machine flow. This is inconvenient when we want to prompt the user to confirm the removal when multiple paths are given. Things are even more complicated when multiple paths are of different types: some paths are active eden checkouts while some are non-eden directories. (**metrics to be added to track this in the future**)

Python version of the remove command is not handling this case well. It will only prompt the user for the first non-eden path if any is given:
 {F1974226015}

## This Diff

We get rid of `State` enum and use functions instead.

The `State` enum was chosen to match [the flow designed as a state machine](https://docs.google.com/document/d/1iAzODvfRuUpu6vwgpEp_82Z0LV2KDYNJyvTILnZUyTE/edit?usp=sharing). However, as the development iterates, we realized that the enum implementation is causing overhead when refactoring code and not carrying any benefits since we don't store values as state data. Switching to functions makes the code easily maintainable. And the state machine flow still exits, just represented by a different implementation.

With this change, the follow-up change to enable multi-paths support becomes fairly straightforward.

## Nest Step

With the new method implemented, we can decouple the validation/determination steps from the state machine flow. The designed flow for multiple paths would be:
1. ~~For each path given, determine the type~~
2. ~~Prompt the user once we have all the paths determined~~
3. After confirmation is given, iterate through the paths and start the state machine for each of them

Reviewed By: MichaelCuevas

Differential Revision: D67928551

fbshipit-source-id: d1c2021d06447911ee33acaaaf6b8987b20e0a71