Bump if counter returns error
Summary:
According to https://fb.prod.workplace.com/groups/1989752134610829/permalink/3773211902931501/ fetching the counter is eventually consistent, and might return an error if it's only been recently created. We will often hit this case as new counter swill have to be created for each client_main_id.

This diff stopps these errors from being logged to scuba to avoid retention issues, and bumps the counter to ensure a value is set for accuracy for the next request

Differential Revision: D68157804

fbshipit-source-id: 09a6a9c73f352fa040af4d4fd250fe4f07510d60