Integration test highlighting lack of support for refs pointing to standalone nested trees
Summary:
Although we now do support importing refs pointing to trees and blobs, we just import the direct objects (tree or blob) themselves. If the ref points to a root tree and there are no other refs that indirectly point to this tree via a commit, then the gitimport mechanism is expected to recursively import every tree and blob in the hierarchy of the root tree. However, currently we just import the top level tree object and move on.

This shortcoming is highlighted by this integration test.

Differential Revision: D68207355

fbshipit-source-id: edba64ed929d199a95bb99217c23ae2114399cba