Implement support for refs pointing to standalone nested trees
Summary:
As discussed in the previous diff, this is missing from the current implementation. This diff adds the necessary changes to support this scenario. Note that this will probably incur a heavy overhead in case the subtrees of the root tree are already uploaded, but that's fine.

The support being added in this diff (and in future diffs) is just to main feature parity with OSS Git. Mononoke Git will NOT advocate for refs pointing to blobs and trees because there is no reason to have those. For all such purposes, committing the changes and pointing a branch/tag to that commit would be the recommended way to go.

Differential Revision: D68208250

fbshipit-source-id: 74a6b170b9696d0d5d665ffdef46313ba642074e